requirements.txt
asgiref==3.5.2
Django==4.1.4
django-jazzmin==2.6.0
djangorestframework==3.14.0
mysqlclient==2.1.1
Pillow==9.3.0
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
